requirements.txt
django==4.1.5
pillow==9.4.0
psycopg2-binary==2.9.5
pytz==2022.7
python-dotenv==0.21.0
django-phonenumbers==1.0.1
django-phonenumber-field==7.0.1
django-ckeditor==6.5.1
django-compressor==4.3
django-crispy-forms==1.14.0
crispy-tailwind==0.5.0
django-modeltranslation==0.18.7
gunicorn==20.1.0
